refactor(util): move file related to file lib
