+++ packages/file/.eslintrc.json
@@ -21,7 +21,7 @@
                 "@nx/dependency-checks": [
                     "error",
                     {
-                        "ignoredDependencies": []
+                        "ignoredDependencies": ["@/log", "@/util"]
                     }
                 ]
             }
