+++ packages/file/package.json
@@ -2,7 +2,8 @@
     "name": "@/file",
     "version": "0.0.1",
     "dependencies": {
-        "tslib": "^2.3.0"
+        "tslib": "^2.3.0",
+        "@nx/devkit": "17.0.0"
     },
     "private": true,
     "type": "commonjs",
