+++ packages/util/src/index.ts
@@ -4,9 +4,6 @@ export { getProjectName } from './lib/getProjectName/getProjectName';
 export { getProjectConfigurationType } from './lib/getProjectConfigurationType/getProjectConfigurationType';
 export { getProjectTypeFromConfig } from './lib/getProjectTypeFromConfig/getProjectTypeFromConfig';
 export { getProjectFromFile } from './lib/getProjectFromFile/getProjectFromFile';
-export { readFileWithTree } from './lib/readFileWithTree/readFileWithTree';
-export { updateFile } from './lib/updateFile/updateFileFile';
-export { writeFileWithTree } from './lib/writeFileWithTree/writeFileWithTree';
 export { getWorkspaceLayout } from './lib/getWorkspaceLayout/getWorkspaceLayout';
 export { runCommandFromDirectory } from './lib/runCommandFromDirectory/runCommandFromDirectory';
 export { snakeCaseToCamelCase } from './lib/snakeCaseToCamelCase/snakeCaseToCamelCase';
