+++ plugins/nx-cmake/src/generators/init/generateReadMe/generateReadMe.spec.ts
@@ -1,7 +1,7 @@
 import type { Tree } from '@nx/devkit';
 import { createTreeWithEmptyWorkspace } from '@nx/devkit/testing';
+import { readFileWithTree } from '@/file';
 import { generateReadMe } from './generateReadMe';
-import { readFileWithTree } from '@/util';
 
 describe('generateReadMe', () => {
     let tree: Tree;
